add dataset to workspace when upload from context area section
