+++ applications/fishing-map/src/features/datasets/NewDatasetTooltip.tsx
@@ -4,12 +4,9 @@ import { useTranslation } from 'react-i18next'
 import { useDispatch, useSelector } from 'react-redux'
 import { Spinner } from '@globalfishingwatch/ui-components'
 import { Dataset } from '@globalfishingwatch/api-types/dist'
-import { selectContextAreasDataviews } from 'features/workspace/workspace.selectors'
-import { getContextDataviewInstance } from 'features/dataviews/dataviews.utils'
-import { useDataviewInstancesConnect } from 'features/workspace/workspace.hook'
 import { AsyncReducerStatus } from 'types'
 import { selectUserDatasetsNotUsed } from 'features/user/user.selectors'
-import { useDatasetModalConnect } from './datasets.hook'
+import { useDatasetModalConnect, useNewDatasetConnect } from './datasets.hook'
 import styles from './NewDatasetTooltip.module.css'
 import {
   fetchAllDatasetsThunk,
@@ -21,9 +18,8 @@ function NewDatasetTooltip({ onSelect }: { onSelect?: (dataset: Dataset) => void
   const { t } = useTranslation()
   const dispatch = useDispatch()
   const { dispatchDatasetModal } = useDatasetModalConnect()
-  const { upsertDataviewInstance } = useDataviewInstancesConnect()
+  const { addNewDatasetToWorkspace } = useNewDatasetConnect()
   const datasets = useSelector(selectUserDatasetsNotUsed)
-  const dataviews = useSelector(selectContextAreasDataviews)
   const datasetsStatus = useSelector(selectDatasetsStatus)
   const allDatasetsRequested = useSelector(selectAllDatasetsRequested)
 
@@ -38,10 +34,7 @@ function NewDatasetTooltip({ onSelect }: { onSelect?: (dataset: Dataset) => void
   }
 
   const onSelectClick = async (dataset: any) => {
-    const usedColors = dataviews?.flatMap((dataview) => dataview.config?.color || [])
-
-    const dataviewInstance = getContextDataviewInstance(dataset.id, usedColors)
-    upsertDataviewInstance(dataviewInstance)
+    addNewDatasetToWorkspace(dataset)
     if (onSelect) {
       onSelect(dataset)
     }
