+++ applications/fishing-map/src/features/datasets/datasets.hook.ts
@@ -3,6 +3,9 @@ import { useSelector, useDispatch } from 'react-redux'
 import { useCallback } from 'react'
 import { Dataset } from '@globalfishingwatch/api-types/dist'
 import { AsyncError } from 'utils/async-slice'
+import { selectContextAreasDataviews } from 'features/workspace/workspace.selectors'
+import { getContextDataviewInstance } from 'features/dataviews/dataviews.utils'
+import { useDataviewInstancesConnect } from 'features/workspace/workspace.hook'
 import {
   CreateDataset,
   createDatasetThunk,
@@ -15,6 +18,22 @@ import {
   setEditingDatasetId,
   updateDatasetThunk,
 } from './datasets.slice'
+
+export const useNewDatasetConnect = () => {
+  const dataviews = useSelector(selectContextAreasDataviews)
+  const { upsertDataviewInstance } = useDataviewInstancesConnect()
+  const usedColors = dataviews?.flatMap((dataview) => dataview.config?.color || [])
+
+  const addNewDatasetToWorkspace = useCallback(
+    (dataset) => {
+      const dataviewInstance = getContextDataviewInstance(dataset.id, usedColors)
+      upsertDataviewInstance(dataviewInstance)
+    },
+    [upsertDataviewInstance, usedColors]
+  )
+
+  return { addNewDatasetToWorkspace }
+}
 
 export const useDatasetModalConnect = () => {
   const dispatch = useDispatch()
